Refactored to separate argument handling from `instantiation.py` and optimize parameters.
