+++ instantiation/controller/agent/agent.py
@@ -3,8 +3,8 @@
 
 from typing import Dict, List
 
-from instantiation.controller.prompter.agent_prompter import (FilterPrompter,
-                                                              PrefillPrompter)
+from controller.prompter.agent_prompter import FilterPrompter, PrefillPrompter
+
 from ufo.agents.agent.basic import BasicAgent
