+++ instantiation/controller/workflow/choose_template_flow.py
@@ -4,36 +4,38 @@ import random
 import time
 import warnings
 from datetime import datetime
-from typing import Dict
 from pathlib import Path
+from typing import Dict
 
+from config.config import Config
 from langchain.embeddings import CacheBackedEmbeddings
 from langchain.storage import LocalFileStore
 from langchain_community.embeddings import HuggingFaceEmbeddings
 from langchain_community.vectorstores import FAISS
 
-from instantiation.config.config import Config
-from instantiation.instantiation import AppEnum
+_configs = Config.get_instance().config_data
 
-configs = Config.get_instance().config_data
 
 class ChooseTemplateFlow:
     """
     Class to select and copy the most relevant template file based on the given task context.
     """
+
     _SENTENCE_TRANSFORMERS_PREFIX = "sentence-transformers/"
 
-    def __init__(self, app_object: AppEnum, task_file_name: str):
+    def __init__(self, app_name: str, file_extension: str, task_file_name: str):
         """
         Initialize the flow with the given task context.
-        :param app_object: An instance of AppEnum, representing the application context.
+        :param app_name: The name of the application.
+        :param file_extension: The file extension of the template.
         :param task_file_name: The name of the task file.
         """
-        self._app_object = app_object
+        self._app_name = app_name
+        self._file_extension = file_extension
         self._task_file_name = task_file_name
         self.execution_time = 0
         self._embedding_model = self._load_embedding_model(
-            model_name=configs["CONTROL_FILTER_MODEL_SEMANTIC_NAME"]
+            model_name=_configs["CONTROL_FILTER_MODEL_SEMANTIC_NAME"]
         )
 
     def execute(self) -> str:
@@ -46,7 +48,9 @@ class ChooseTemplateFlow:
         self.execution_time = round(time.time() - start_time, 3)
         return template_copied_path
 
-    def _create_copied_file(self, copy_from_path: Path, copy_to_folder_path: Path, file_name: str = None) -> str:
+    def _create_copied_file(
+        self, copy_from_path: Path, copy_to_folder_path: Path, file_name: str = None
+    ) -> str:
         """
         Create a cache file from the specified source.
         :param copy_from_path: The original path of the file.
@@ -55,9 +59,10 @@ class ChooseTemplateFlow:
         :return: The path to the newly created cache file.
         """
         os.makedirs(copy_to_folder_path, exist_ok=True)
-        copied_template_path = self._generate_copied_file_path(copy_to_folder_path, file_name)
+        copied_template_path = self._generate_copied_file_path(
+            copy_to_folder_path, file_name
+        )
 
-        # Copy the content from the original file to the new file
         with open(copy_from_path, "rb") as f:
             ori_content = f.read()
         with open(copied_template_path, "wb") as f:
@@ -72,7 +77,7 @@ class ChooseTemplateFlow:
         :param file_name: Optional; the name of the task file.
         :return: The path to the newly created file.
         """
-        template_extension = self._app_object.file_extension
+        template_extension = self._file_extension
         if file_name:
             return str(folder_path / f"{file_name}{template_extension}")
         timestamp = datetime.now().strftime("%Y-%m-%d-%H-%M-%S")
@@ -83,14 +88,19 @@ class ChooseTemplateFlow:
         Choose the most relevant template file based on the task.
         :return: The path to the most relevant template file.
         """
-        templates_description_path = Path(configs["TEMPLATE_PATH"]) /\
-              self._app_object.description.lower() / "description.json"
+        templates_description_path = (
+            Path(_configs["TEMPLATE_PATH"]) / self._app_name / "description.json"
+        )
 
         try:
             with open(templates_description_path, "r") as f:
-                return self._choose_target_template_file(self._task_file_name, json.load(f))
+                return self._choose_target_template_file(
+                    self._task_file_name, json.load(f)
+                )
         except FileNotFoundError:
-            warnings.warn(f"Warning: {templates_description_path} does not exist. Choosing a random template.")
+            warnings.warn(
+                f"Warning: {templates_description_path} does not exist. Choosing a random template."
+            )
             return self._choose_random_template()
 
     def _choose_random_template(self) -> str:
@@ -98,7 +108,7 @@ class ChooseTemplateFlow:
         Select a random template file from the template folder.
         :return: The path to the randomly selected template file.
         """
-        template_folder = Path(configs["TEMPLATE_PATH"]) / self._app_object.description.lower()
+        template_folder = Path(_configs["TEMPLATE_PATH"]) / self._app_name
         template_files = [f for f in template_folder.iterdir() if f.is_file()]
 
         if not template_files:
@@ -114,23 +124,33 @@ class ChooseTemplateFlow:
         :return: The path to the copied template file.
         """
         chosen_template_file_path = self._get_chosen_file_path()
-        chosen_template_full_path = Path(configs["TEMPLATE_PATH"]) / \
-            self._app_object.description.lower() / chosen_template_file_path
+        chosen_template_full_path = (
+            Path(_configs["TEMPLATE_PATH"]) / self._app_name / chosen_template_file_path
+        )
 
-        target_template_folder_path = Path(configs["TASKS_HUB"]) / \
-            (os.path.dirname(os.path.dirname(self._task_file_name)) + "_templates")
+        target_template_folder_path = Path(_configs["TASKS_HUB"]) / (
+            os.path.dirname(os.path.dirname(self._task_file_name)) + "_templates"
+        )
 
-        return self._create_copied_file(chosen_template_full_path, target_template_folder_path, self._task_file_name)
+        return self._create_copied_file(
+            chosen_template_full_path, target_template_folder_path, self._task_file_name
+        )
 
-    def _choose_target_template_file(self, given_task: str, doc_files_description: Dict[str, str]) -> str:
+    def _choose_target_template_file(
+        self, given_task: str, doc_files_description: Dict[str, str]
+    ) -> str:
         """
         Get the target file based on the semantic similarity of the given task and the template file descriptions.
         :param given_task: The task to be matched.
         :param doc_files_description: A dictionary of template file descriptions.
         :return: The path to the chosen template file.
         """
-        file_doc_map = {desc: file_name for file_name, desc in doc_files_description.items()}
-        db = FAISS.from_texts(list(doc_files_description.values()), self._embedding_model)
+        file_doc_map = {
+            desc: file_name for file_name, desc in doc_files_description.items()
+        }
+        db = FAISS.from_texts(
+            list(doc_files_description.values()), self._embedding_model
+        )
         most_similar = db.similarity_search(given_task, k=1)
 
         if not most_similar:
@@ -145,8 +165,10 @@ class ChooseTemplateFlow:
         :param model_name: The name of the embedding model to load.
         :return: The loaded embedding model.
         """
-        store = LocalFileStore(configs["CONTROL_EMBEDDING_CACHE_PATH"])
+        store = LocalFileStore(_configs["CONTROL_EMBEDDING_CACHE_PATH"])
         if not model_name.startswith(ChooseTemplateFlow._SENTENCE_TRANSFORMERS_PREFIX):
             model_name = ChooseTemplateFlow._SENTENCE_TRANSFORMERS_PREFIX + model_name
         embedding_model = HuggingFaceEmbeddings(model_name=model_name)
-        return CacheBackedEmbeddings.from_bytes_store(embedding_model, store, namespace=model_name)
+        return CacheBackedEmbeddings.from_bytes_store(
+            embedding_model, store, namespace=model_name
+        )
